Soft constraints to separate file
